move ts stuff to own files
